requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.1
idna==3.11
packaging==26.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
PyJWT==2.8.0
